+++ src/main/java/com/example/hrm/dto/response/DepartmentResponse.java
@@ -1,17 +1,17 @@
 package com.example.hrm.dto.response;
 
 public class DepartmentResponse {
-    private int id;
+    private Long id;
     private String name;
     private int quantity;
 
-    public DepartmentResponse(int id,String name, int quantity) {
+    public DepartmentResponse(Long id,String name, int quantity) {
         this.id = id;
         this.name = name;
         this.quantity = quantity;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
